Divisão com numeros decimais

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -34,7 +34,7 @@ public class Calculadora {
 		return a * b;
 	}
 
-	public static int divisao(int a, int b) {
+	public static float divisao(int a, int b) {
 		return a / b;
 	}
